Création de posts à l'aide d'une transaction

+++ Post/index.php
@@ -1,6 +1,8 @@
 <?php
 require_once "../php/sql_func.inc.php";
 
+$commentairePost = filter_input(INPUT_POST,'postTextArea',FILTER_SANITIZE_STRING);
+
 if(!isset($_SESSION))
   session_start();
 
@@ -35,12 +37,13 @@ $_SESSION['currentPage'] =  "Post";
 <form action="./index.php" method="POST" enctype="multipart/form-data"> 
 
 <label for="postTextArea">Entrez du text</label></br>
-<textarea required id="postTextArea" rows="3" cols="50"></textarea></br>
+<textarea required name="postTextArea" id="postTextArea" rows="3" cols="50"></textarea></br>
 <label for="fileSelect"> Select a file:</label> <input id="fileSelect" accept="image/*" type="file" name="imgSelect[]" multiple>
 <input type="submit">
 </form>
 
 <?php 
+$UserPostImages = [];
 for($i = 0; $i < count($_FILES["imgSelect"]['name']) ; $i++)
 {
   $Orgfilename = $_FILES["imgSelect"]["name"][$i];
@@ -50,25 +53,62 @@ for($i = 0; $i < count($_FILES["imgSelect"]['name']) ; $i++)
   $ext = explode("/",$_FILES["imgSelect"]["type"][$i])[1];
   $file = $filename.'.'.$ext;
   
-    if(in_array($ext,["png","bmp","jpg","jpeg","gif"]) && $_FILES["imgSelect"]['size'][$i] < 3145728)
+
+  if(in_array($ext,["png","bmp","jpg","jpeg","gif"]) && $_FILES["imgSelect"]['size'][$i] < 3145728)
+  {
+    if($commentairePost != "")
     {
-      if(move_uploaded_file($_FILES["imgSelect"]["tmp_name"][$i],$dir.$file))
-      {
-        
-        if(uploadImg($filename,$ext))
-        echo "Fichiers uploadés";
-        else
-        {
-          echo "Error lors de l'upload";
-          unlink($dir.$file);
-        }   
-      }
-      else
-      echo "Error lors de l'upload";
+      array_push($UserPostImages, [$filename,$ext]);
+    
+
+     
     }
     else
+    {
+      echo "Veuillez écrire un commentaire";
+      return;
+    }
+  
+    
+  }
+  else
+  {
     echo "Veuillez sélectionner des fichiers valides!";
+    return;
+  }
+
+
+  
+
+/*
+    if(in_array($ext,["png","bmp","jpg","jpeg","gif"]) && $_FILES["imgSelect"]['size'][$i] < 3145728)
+    {
+      if($commentairePost != "")
+      {
+        if(move_uploaded_file($_FILES["imgSelect"]["tmp_name"][$i],$dir.$file))
+        {       
+          if(uploadImg($filename,$ext))
+          echo "Fichiers uploadés";
+          else
+          {
+            echo "Error lors de l'upload";
+            unlink($dir.$file);
+          }   
+        }
+        else
+        echo "Error lors de l'upload";
+      }
+      else
+      echo "Veuillez écrire un commentaire";
+      
+    }
+    else
+    echo "Veuillez sélectionner des fichiers valides!";*/
 }
+if(createNewPost($commentairePost,$UserPostImages))
+  echo "Post Crée";
+  else
+  echo "Erreur lors de la création du Post";
 var_dump($ext);
 var_dump($_FILES["imgSelect"]);
 ?>
